lib/database: make the Iterator interface portable
The NewIterator method on Database returned an Iterable, which was
actually tied to the BadgerDB implementation. Rename Iterable to
BadgerIterator, and make NewIterator return the Iterator interface.

This requires a few type assertions in the tests, but the tests already
assume the BadgerDB implementation in those places, so that's fine for
now.

While at it, consolidate the Database interface with the Reader and
Writer interfaces. This is easier to maintain and clearer, as there are
fewer duplicated methods.