Added checks to avoid decoding messages as handshakes (#2681)
Sometimes, we end up decoding a message as a handshake and IsHandshake is not useful after decoding has happened. 

IsHandshake in Handshake implementations has been improved to check if the value of `Roles` is valid. If we decode bytes of handshake to handshake Roles would have a valid value (1,2, 4) otherwise it would have some random value.