Author a sibling block in case best block's slot is same as current slot  (#2827)
While authoring a block in a given slot S, it could happen that best block was also authored in the slot S by some other author. In this case, we can't create a new block as a child of best block, because then we would have a chain where slot numbers don't increase.
In such scenario, we author a sibling block to best block. That is we author a block as a child of best block's parent.
After authoring such a block, we would have two blocks with same parent, same block number and same slot. This would create a fork, but that would eventually get resolved.