enhancement(pkg/scale): remove sorting logic from encodeMap method (#2904)
In go it's understood that the order of the keys is non-deterministic. In this case, we should be testing that we can encode a map and decode into a map, and assert that the maps are equal. 
This commit ensures that decoding into a rust BTreeMap via the parity-scale-codec works as expected. To make the tests more deterministic, we have taken a map with a single key or a map with two keys, and asserted that it's equal to one of the two encoding and removed the sorting logic altogether.