check hash on verifying block justification
So far, while verifying block justification in grandpa, we were just
checking if we already know a finalised block that has the same set id and
round as the block in question.

we should check if the finalised block that we know already and the
block that we are verifying are not the same and not throw error in this
case.

Fixes #3061