fix(lib/grandpa):  on verifying block justification, compare given block hash with finalised hash (#3081)
So far, while verifying block justification in grandpa, we were just checking if we already know a finalised block that has the same set id and round as the block in question.

We should check if the finalised block that we know already and the block that we are verifying are not the same, and not throw error in this case.

We should through error if we see two finalised blocks with the same set id and round. In this case, both blocks that we were comparing were the same. We were not comparing their hashes and were throwing an error thinking that we have two finalised blocks in the same set id and round.

After this change, I was able to sync further than 198656.