chore(dot/network): added common request-response protocol (#3334)
So far, we were using only one request-response protocol (block request and block response). But now, we need to add quite a few more. Just like we did for notification protocol, we need some common code-infrastructure for adding new request-response protocol. 

This PR introduces interfaces for Request and Response to follow and common functions to make request and decode response.