fix(dot): use tempDir in tests as base path to avoid creating `dot/~` (#3363)
On running integration tests in for dot say go test -tags integration --run github.com/ChainSafe/gossamer/dot/... -v, a folder named `dot/~` was getting created. We shouldn't create a folder named '~'. 

This was happening because basepath was starting with '~', as '~' was supposed to be parsed into home later. This commit changes basepath to a temp directory in all tests.