fix: add a limit of number of bytes while scale decoding a slice
While scale decoding we first read the length of bytes to decode and
then we decode that many bytes.

Someone could ask us to decode some malicious bytes such that the read
length is unreasonably big. In such case, we would have to create a byte
slice as big as the length. So, we would either panic because
of `makeslice: len out of range` or because the asked length would be
more than the memory we have available in our machine.

Both of these are the situations we don't want. Hence, I have put a
limit on decodable bytes (maxDecodableBytes), 16MB. I have kept it 16MB
because for the usecase of polkadot specs, we would never need to decode
a bigger set of bytes.