fix: add a limit of number of bytes while scale decoding a slice (#3733)
While scale decoding we first read the length of bytes to decode and
then we decode that many bytes.

Someone could ask us to decode some malicious bytes such that the read
length is unreasonably big. In such case, we would have to create a byte
slice as big as the length. The length in byte slice is an encoded as `Compact<u32>`. 

Current we are just reading length as uint and not checking if it goes beyond the bounds of uint32.
So, we would either panic because of `makeslice: len out of range` or because the asked length would be
more than the memory we have available in our machine.

We are going to put a check to makes sure that this length is less than max of uint32.